+++ Cache/ICachePage.cs
@@ -9,7 +9,7 @@
         bool IsReady();
         float GetElevation(int x, int y);
         float GetDetail(int x, int y);
-        ushort GetTree(int x, int y);
+        int GetTree(int x, int y);
         Vector3 GetPosition(int x, int y);
         Vector3 GetNormal(int x, int y);
         Color3 GetColor(int x, int y);
